First attempt. Does not work.
I think something is wrong in the RHS. I have checked the matrix
thoroughly and I am pretty confident that it is correct. I am not so
confident about the rhs. In any event, I dont think this is going to
get me any computational gains (even using the scipy iterative
solvers).